==> analysis/01_dataset_creation.Rmd <==
---
title: "SID AKI Study: dataset creation"
author: Miguel Ángel Armengol de la Hoz
date: '`r format(Sys.Date(), "%B %d, %Y")`'
output:
  html_notebook:
    code_folding: show
    number_sections: yes
    theme: flatly
    toc: yes
    toc_float: yes

knit: (function(inputFile, encoding) { rmarkdown::render(inputFile, encoding = encoding, output_file = paste0(substr(inputFile,1,nchar(inputFile)-4)," ",Sys.Date(),'.html')) })
---

# Environment

```{r message=FALSE, warning=FALSE}
library(bigrquery)
library(summarytools)
library(readr)
library(stringr)
library(sqldf)
library(dplyr)
library(readxl)
```


# Set up BigQuery related functions

This chunks also creates the run_query and get_sql function.

```{r setup, include=FALSE}

project_id <- "hst-953-2018"
options(httr_oauth_cache=FALSE)
run_query <- function(query){
  data <- query_exec(query, project=project_id, use_legacy_sql=FALSE,max_pages = Inf)
  return(data)
}

# function for reading sql files
getSQL <- function(filepath){
  con = file(filepath, "r")
  sql.string <- ""

  while (TRUE){
    line <- readLines(con, n = 1)

    if ( length(line) == 0 ){
      break
    }

    line <- gsub("\\t", " ", line)

    if(grepl("--",line) == TRUE){
      line <- paste(sub("--","/*",line),"*/")
    }

    sql.string <- paste(sql.string, line)
  }

  close(con)
  return(sql.string)
}
```

# Loading queries and extracting the data

Loads all queries from the sql files in the extraction folder and runs them into RBigQuey to extract the data.

```{r}
#folders contained in sql folder
#cat(list.files('sql'), sep = '\n')

# AKI related variables
first_creat<-run_query(getSQL("sql/aki/first_creat.sql"))
before_adm_creat<-run_query(getSQL("sql/aki/before_adm_creat.sql"))
adm_creat<-run_query(getSQL("sql/aki/adm_creat.sql"))
chronicAKI<-run_query(getSQL("sql/aki/chronicAKI.sql"))
first_rrt<-run_query(getSQL("sql/aki/first_rrt.sql"))
peakcreat48h<-run_query(getSQL("sql/aki/peakcreat48h.sql"))
peakcreat7days<-run_query(getSQL("sql/aki/peakcreat7days.sql"))

dialysis<-run_query(getSQL("sql/aki/dialysis.sql"))
dialysis_first7days<-run_query(getSQL("sql/aki/dialysis_first7days.sql"))


# rest of the variables
bg_24_48<-run_query(getSQL("sql/bg_24_48.sql"))
bg_first<-run_query(getSQL("sql/bg_first.sql"))
demographics<-run_query(getSQL("sql/demographics.sql"))
lab_first<-run_query(getSQL("sql/lab_first.sql"))
lab_24<-run_query(getSQL("sql/lab_24.sql"))
lab_48<-run_query(getSQL("sql/lab_48.sql"))
outcomes<-run_query(getSQL("sql/outcomes.sql"))
pasthistory<-run_query(getSQL("sql/pasthistory.sql"))
sid_first<-run_query(getSQL("sql/sid_first.sql"))
sid_24_48<-run_query(getSQL("sql/sid_24_48.sql"))
sid_twa_24<-run_query(getSQL("sql/sid_twa_24.sql"))
sid_twa_48<-run_query(getSQL("sql/sid_twa_48.sql"))
temp_day1<-run_query(getSQL("sql/temp_day1.sql"))
temp_day2<-run_query(getSQL("sql/temp_day2.sql"))
urine_output_24_48<-run_query(getSQL("sql/urine_output_24_48.sql"))
vent_bin_first_2days<-run_query(getSQL("sql/vent_bin_first_2days.sql"))
vitals_day1<-run_query(getSQL("sql/vitals_day1.sql"))
vitals_day2<-run_query(getSQL("sql/vitals_day2.sql"))


```

## AKI Dataset creation

```{r}
#We need to exclude patients who already had AKI before ICU admin)
AKIdevelopment<-subset(baseline_creat, !(baseline_creat$patientunitstayid %in% chronicAKI$patientunitstayid) )
AKIdevelopment<-left_join(AKIdevelopment, peakcreat7days)
AKIdevelopment<-left_join(AKIdevelopment, peakcreat48h)
AKIdevelopment$position<-NULL
```

### Windsorization

Clean data by means of winsorization, i.e., by shrinking outlying observations to the border of the main part of the data.
The creatinine values have been winsorize replacing the extreme observations using 95% and 5% percentiles.

```{r}
AKIdevelopmentW<-AKIdevelopment
## Winsorization: Replace the extreme observations using 95% and 5% percentiles
winsorize_x = function(x, cut = 0.05){
  cut_point_top <- quantile(x, 1 - cut, na.rm = T)
  cut_point_bottom <- quantile(x, cut, na.rm = T)
  i = which(x >= cut_point_top) 
  x[i] = cut_point_top
  j = which(x <= cut_point_bottom) 
  x[j] = cut_point_bottom
  return(x)
}

AKIdevelopmentW$creat1<-winsorize_x(AKIdevelopmentW$creat1)
AKIdevelopmentW$peakcreat7d<-winsorize_x(AKIdevelopmentW$peakcreat7d)
AKIdevelopmentW$peakcreat48h<-winsorize_x(AKIdevelopmentW$peakcreat48h)

#Now we turn all emptpy values into 0

AKIdevelopmentW$creat1[is.na(AKIdevelopmentW$creat1)]<-0
AKIdevelopmentW$peakcreat7d[is.na(AKIdevelopmentW$peakcreat7d)]<-0
AKIdevelopmentW$peakcreat48h[is.na(AKIdevelopmentW$peakcreat48h)]<-0


```

###List of patients with AKI and offset

```{r}

AKIdevelopmentW['cr7undercreat1']<-AKIdevelopmentW$peakcreat7d/AKIdevelopmentW$creat1
AKIdevelopmentW['cr48hdundercreat1']<-AKIdevelopmentW$peakcreat48h-AKIdevelopmentW$creat1
AKIdevelopmentW['AKIoffset']<-NA
AKIdevelopmentW['AKIstage']<-NA

#nested if else function

i <- function(if_stat, then) {
  if_stat <- lazyeval::expr_text(if_stat)
  then    <- lazyeval::expr_text(then)
  sprintf("ifelse(%s, %s, ", if_stat, then)
}

e <- function(else_ret) {
  else_ret <- lazyeval::expr_text(else_ret)
  else_ret
}

ie <- function(...) {
  args <- list(...)
  
  for (i in 1:(length(args) - 1) ) {
      if (substr(args[[i]], 1, 6) != "ifelse") {
        stop("All but the last argument, need to be i functions.", call. = FALSE)
      }
  }
  if (substr(args[[length(args)]], 1, 6) == "ifelse"){
    stop("Last argument needs to be an e function.", call. = FALSE)
  }
  args$final <- paste(rep(')', length(args) - 1), collapse = '')
  eval_string <- do.call('paste', args)
  eval(parse(text = eval_string))
}

AKIdevelopmentW$AKIstage <- 
  ie(
    i(AKIdevelopmentW$cr7undercreat1>=3,   3),
    i(AKIdevelopmentW$cr7undercreat1>=2,   2),
    i(AKIdevelopmentW$cr7undercreat1>=1.5,   1),
    i(AKIdevelopmentW$cr48hdundercreat1>=0.3,   1),
    e(0)
  )

AKIdevelopmentW$AKIoffset <- 
  ie(
    i(AKIdevelopmentW$cr7undercreat1>=3,AKIdevelopmentW$peakcreat7d_offset),
    i(AKIdevelopmentW$cr7undercreat1>=2,AKIdevelopmentW$peakcreat7d_offset),
    i(AKIdevelopmentW$cr7undercreat1>=1.5,AKIdevelopmentW$peakcreat7d_offset),
    i(AKIdevelopmentW$cr48hdundercreat1>=0.3,   AKIdevelopmentW$peakcreat48h_offset),
    e(0)
  )

library(dplyr)
AKIlist<-AKIdevelopmentW %>%
  select(patientunitstayid,AKIstage
         # ,AKIoffset The nephrologist suggested not to use this variable since we can not infer when AKI was developed from it. Removing it to avoid confusion.
         )
```

### Final List of patients with AKI and offset (with RRT)

Offset is not reliable so we exclude it.

```{r}
# Dr Danziger said the offset is not reliable for telling when the patient exactly developed AKI so he adviced not to use it

AKIlist_final<-left_join(AKIlist,first_rrt)


AKIlist_final$AKIstage <- 
  ie(
    i(!is.na(AKIlist_final$first_rrtoffset),   3),
    e(AKIlist_final$AKIstage)
  )

# The nephrologist suggested not to use this variable since we can not infer when AKI was developed from it. Removing it to avoid confusion.
# AKIlist_final$AKIoffset <- 
#   ie(
#     i(!is.na(AKIlist_final$first_rrtoffset), AKIlist_final$first_rrtoffset),
#     e(AKIlist_final$AKIoffset)
#   )

AKIlist_final$first_rrtoffset<-NULL
AKIlist_final[is.na(AKIlist_final)]<-0
AKIlist_final$rrt_bin<-NULL
```

# New Variables

```{r}
admission_vars<-run_query(getSQL("sql/admission_vars.sql"))
mean_first_day<-run_query(getSQL("sql/mean_first_day.sql"))
total_na_cl_volume_day1<-run_query(getSQL("sql/total_na_cl_volume_day1.sql"))
intake_day1_missing_explain<-run_query(getSQL("sql/intake_day1_missing_explain.sql"))
reliable_fluids_data<-run_query(getSQL("sql/all_reliable.sql"))

```

# Merging all datasets

We are using a left join to join them

```{r}
# Exclude non reliable data ICU-year
sid_aki_study<-inner_join(demographics,reliable_fluids_data)

# left join = all.x=TRUE
sid_aki_study<-Reduce(function(...) merge(..., all.x=TRUE), list(
  sid_aki_study
  
, bg_first
, bg_24_48

, lab_first
, lab_24
, lab_48

, outcomes
, pasthistory
, sid_first
, sid_24_48
, sid_twa_24
, sid_twa_48
, temp_day1
, temp_day2
, urine_output_24_48

, vent_bin_first_2days

, vitals_day1
, vitals_day2
# aki vars
, first_creat
, before_adm_creat
, adm_creat
, chronicAKI
, first_rrt
, peakcreat48h
, peakcreat7days
, AKIlist_final

# new variables requested
, admission_vars
, mean_first_day
, total_na_cl_volume_day1
, intake_day1_missing_explain
, dialysis_first7days

))


#cat(names(sid_aki_study),sep = '\n')
```

According to the attached file, we have 12,551 patients with fluid data after inclusion and exclusion criteria were applied.


If fluid data=NA, appears_in_intakeoutput=0, receives_intake_atsomepoint=0, then I think those are missing value (n=680).

If fluid data=NA, appears_in_intakeoutput=1; receives_intake_atsomepoint=0, then I guess those patients took no fluid intake (n=4,733).  ← That could be too many...


If fluid data=NA, appears_in_intakeoutput=1, receives_intake_atsomepoint=1, I do not know if those are missing value or from our mistake (n=13,568). The number of patients are too many, so we should check further. In other words, we could miss important fluids in this table.
https://docs.google.com/spreadsheets/d/17BtBk7zkSqgUxGriTbJtJt0VrFKpnBQxLVyywR_iHKM/edit#gid=100723980  

Please look at those 13,568 patients and check if there are some variables that have large amount of cellvaluenumeric, let's say, more than 500 ml. That could be important fluid that we missed. If you find that kind of data, please send them to me.

Best,


# Dataset report

```{r}
view(dfSummary(sid_aki_study))
view(dfSummary())
```

# Export dataset

```{r}
write.csv(sid_aki_study,'sid_aki_study.csv',row.names = F)
```
