==> analysis/02_data_wrangling.Rmd <==
---
title: "02"
output: html_document
---

# Environment

```{r}
library(dplyr)
library(Amelia)
```

# Creating new variables

````{r,results='hide'}
#For outcomes/predictors with substantial missig data, IPW should be used?
#Exclude CKD, rrt before admision
df_1<-sid_aki_study%>%
  filter(creat1<=3)%>%
  mutate(aki=case_when(.$AKIstage==0~0,.$AKIstage%in%c(1,2,3)~1))%>%
  mutate(gender=case_when(.$gender=="Female"~0,
                          .$gender=="Male"~1,
                          TRUE~as.numeric(NA)),
         pre_rrt=case_when(.$rrt_bin==1&first_rrtoffset<=0~1,
                          TRUE~0))%>%filter(pre_rrt==0)

df_1$ethnicity[df_1$ethnicity ==""| 
             df_1$ethnicity == "Native American"|
             df_1$ethnicity == "Other/Unknow"] <- "Other/Unknown"

df_1$surgery<-NA
df_1$surgery[df_1$hospitalAdmitSource =="Operating Room"| 
             df_1$hospitalAdmitSource == "PACU"] <- 1
df_1$surgery[df_1$hospitalAdmitSource =="Acute Care/Floor"| 
             df_1$hospitalAdmitSource == "Chest Pain Center"|
             df_1$hospitalAdmitSource == "Direct Admit"|
               df_1$hospitalAdmitSource == "Emergency Department"|
               df_1$hospitalAdmitSource == "Floor"|
               df_1$hospitalAdmitSource == "ICU"|
               df_1$hospitalAdmitSource == "ICU to SDU"|
               df_1$hospitalAdmitSource == "Observation"|
               df_1$hospitalAdmitSource == "Other"|
               df_1$hospitalAdmitSource == "Other Hospital"|
               df_1$hospitalAdmitSource == "Other ICU"|
               df_1$hospitalAdmitSource == "Recovery Room"|
               df_1$hospitalAdmitSource == "Step-Down Unit (SDU)"] <- 0

#Categorizing Na, Cl, and SID
df_2<-df_1%>%mutate(cl_adm_cat=case_when(.$chloride_first<98 ~ 1,
                              .$chloride_first>=98 & .$chloride_first<=110 ~ 0,
                              .$chloride_first>110 ~ 2,
                                TRUE ~ as.numeric(NA)),
         na_adm_cat=case_when(.$sodium_first>145 ~ 2,
                                .$sodium_first<=145  & .$sodium_first>=135 ~ 0,
                                .$sodium_first<135 ~ 1,
                                TRUE ~ as.numeric(NA)),
         sid_adm_cat=case_when(.$SID_first>37 ~ 2,
                                .$SID_first<=37  & .$SID_first>=31 ~ 0,
                                .$SID_first<31 ~ 1,
                                TRUE ~ as.numeric(NA)),
         
         # 24 hours variables start here
         cl_24_min_cat=case_when(.$chloride_min_24<98 ~ 1,
                              .$chloride_min_24>=98 & .$chloride_min_24<=110 ~ 0,
                              .$chloride_min_24>110 ~ 2,
                                TRUE ~ as.numeric(NA)),
         cl_24_max_cat=case_when(.$chloride_max_24<98 ~ 1,
                              .$chloride_max_24>=98 & .$chloride_max_24<=110 ~ 0,
                              .$chloride_max_24>110 ~ 2,
                                TRUE ~ as.numeric(NA)),
         na_24_min_cat=case_when(.$sodium_min_24>145 ~ 2,
                                .$sodium_min_24<=145  & .$sodium_min_24>=135 ~ 0,
                                .$sodium_min_24<135 ~ 1,
                                TRUE ~ as.numeric(NA)),
         na_24_max_cat=case_when(.$sodium_max_24>145 ~ 2,
                                .$sodium_max_24<=145  & .$sodium_max_24>=135 ~ 0,
                                .$sodium_max_24<135 ~ 1,
                                TRUE ~ as.numeric(NA)),
         sid_24_min_cat=case_when(.$SID_min_24>37 ~ 2,
                                .$SID_min_24<=37  & .$SID_min_24>=31 ~ 0,
                                .$SID_min_24<3 ~ 1,
                                TRUE ~ as.numeric(NA)),
         sid_24_max_cat=case_when(.$SID_max_24>37 ~ 2,
                                .$SID_max_24<=37  & .$SID_max_24>=31 ~ 0,
                                .$SID_max_24<31 ~ 1,
                                TRUE ~ as.numeric(NA)),
         cl_24_mean_cat=case_when(.$chloride_twa_24<98 ~ 1,
                              .$chloride_twa_24>=98 & .$chloride_twa_24<=110 ~ 0,
                              .$chloride_twa_24>110 ~ 2,
                                TRUE ~ as.numeric(NA)),
         cl_24_mean_catd=case_when(.$chloride_twa_24>116 ~ 13,
                                   .$chloride_twa_24>114&.$chloride_twa_24<=116 ~ 12,
                                   .$chloride_twa_24>112&.$chloride_twa_24<=114 ~ 11,
                                   .$chloride_twa_24>110&.$chloride_twa_24<=112 ~ 10,
                                   .$chloride_twa_24>108&.$chloride_twa_24<=110 ~ 9,
                                   .$chloride_twa_24>106&.$chloride_twa_24<=108 ~ 8,
                                   .$chloride_twa_24>104&.$chloride_twa_24<=106 ~ 7,
                                   .$chloride_twa_24>102&.$chloride_twa_24<=104 ~ 6,
                                   .$chloride_twa_24>100&.$chloride_twa_24<=102 ~ 5,
                                   .$chloride_twa_24>98&.$chloride_twa_24<=100~ 4,
                                   .$chloride_twa_24>96&.$chloride_twa_24<=98 ~ 3,
                                   .$chloride_twa_24>94&.$chloride_twa_24<=96 ~ 2,
                                   .$chloride_twa_24<=94~1,
                                   TRUE ~ as.numeric(NA)),
         na_24_mean_cat=case_when(.$sodium_twa_24>145 ~ 2,
                                .$sodium_twa_24<=145  & .$sodium_twa_24>=135 ~ 0,
                                .$sodium_twa_24<135 ~ 1,
                                TRUE ~ as.numeric(NA)),
         na_24_mean_catd=case_when(.$sodium_twa_24>146 ~ 11,
                                   .$sodium_twa_24>144&.$sodium_twa_24<=146 ~ 10,
                                   .$sodium_twa_24>142&.$sodium_twa_24<=144 ~ 9,
                                   .$sodium_twa_24>140&.$sodium_twa_24<=142 ~ 8,
                                   .$sodium_twa_24>138&.$sodium_twa_24<=140 ~ 7,
                                   .$sodium_twa_24>136&.$sodium_twa_24<=138 ~ 6,
                                   .$sodium_twa_24>134&.$sodium_twa_24<=136 ~ 5,
                                   .$sodium_twa_24>132&.$sodium_twa_24<=134 ~ 4,
                                   .$sodium_twa_24>130&.$sodium_twa_24<=132 ~ 3,
                                   .$chloride_twa_24<=130~1,
                                   TRUE ~ as.numeric(NA)),
         sid_24_mean_cat=case_when(.$SID_twa_24>37 ~ 2,
                                .$SID_twa_24<=37  & .$SID_twa_24>=31 ~ 0,
                                .$SID_twa_24<31 ~ 1,
                                TRUE ~ as.numeric(NA)),
         sid_24_mean_catd=case_when(.$SID_twa_24>44 ~ 11,
                                .$SID_twa_24>42  & .$SID_twa_24<=44 ~ 10,
                                .$SID_twa_24>40  & .$SID_twa_24<=42 ~ 9,
                                .$SID_twa_24>38  & .$SID_twa_24<=40 ~ 8,
                                .$SID_twa_24>36  & .$SID_twa_24<=38 ~ 7,
                                .$SID_twa_24>34  & .$SID_twa_24<=36 ~ 6,
                                .$SID_twa_24>32  & .$SID_twa_24<=34 ~ 5,
                                .$SID_twa_24>30  & .$SID_twa_24<=32 ~ 4,
                                .$SID_twa_24>28  & .$SID_twa_24<=30 ~ 3,
                                .$SID_twa_24>26  & .$SID_twa_24<=28 ~ 2,
                                .$SID_twa_24<=26 ~ 1,
                                TRUE ~ as.numeric(NA)),
         cl_24_mean_catq=cut(chloride_twa_24,breaks=10,labels=c("1","2","3","4","5","6","7","8","9","10")),
         delta_cl_d1=chloride_twa_24-chloride_first,
         delta_na_d1=sodium_twa_24-sodium_first,
         delta_sid_d1=SID_twa_24-SID_first,
         
         # 48 hours variables start here
          cl_48_min_cat=case_when(.$chloride_min_48<98 ~ 1,
                              .$chloride_min_48>=98 & .$chloride_min_48<=110 ~ 0,
                              .$chloride_min_48>110 ~ 2,
                                TRUE ~ as.numeric(NA)),
         cl_48_max_cat=case_when(.$chloride_max_48<98 ~ 1,
                              .$chloride_max_48>=98 & .$chloride_max_48<=110 ~ 0,
                              .$chloride_max_48>110 ~ 2,
                                TRUE ~ as.numeric(NA)),
         na_48_min_cat=case_when(.$sodium_min_48>145 ~ 2,
                                .$sodium_min_48<=145  & .$sodium_min_48>=135 ~ 0,
                                .$sodium_min_48<135 ~ 1,
                                TRUE ~ as.numeric(NA)),
         na_48_max_cat=case_when(.$sodium_max_48>145 ~ 2,
                                .$sodium_max_48<=145  & .$sodium_max_48>=135 ~ 0,
                                .$sodium_max_48<135 ~ 1,
                                TRUE ~ as.numeric(NA)),
         sid_48_min_cat=case_when(.$SID_min_48>37 ~ 2,
                                .$SID_min_48<=37  & .$SID_min_48>=31 ~ 0,
                                .$SID_min_48<3 ~ 1,
                                TRUE ~ as.numeric(NA)),
         sid_48_max_cat=case_when(.$SID_max_48>37 ~ 2,
                                .$SID_max_48<=37  & .$SID_max_48>=31 ~ 0,
                                .$SID_max_48<31 ~ 1,
                                TRUE ~ as.numeric(NA)),
         cl_48_mean_cat=case_when(.$chloride_twa_48<98 ~ 1,
                              .$chloride_twa_48>=98 & .$chloride_twa_48<=110 ~ 0,
                              .$chloride_twa_48>110 ~ 2,
                                TRUE ~ as.numeric(NA)),
         cl_48_mean_catd=case_when(.$chloride_twa_48>116 ~ 13,
                                   .$chloride_twa_48>114&.$chloride_twa_48<=116 ~ 12,
                                   .$chloride_twa_48>112&.$chloride_twa_48<=114 ~ 11,
                                   .$chloride_twa_48>110&.$chloride_twa_48<=112 ~ 10,
                                   .$chloride_twa_48>108&.$chloride_twa_48<=110 ~ 9,
                                   .$chloride_twa_48>106&.$chloride_twa_48<=108 ~ 8,
                                   .$chloride_twa_48>104&.$chloride_twa_48<=106 ~ 7,
                                   .$chloride_twa_48>102&.$chloride_twa_48<=104 ~ 6,
                                   .$chloride_twa_48>100&.$chloride_twa_48<=102 ~ 5,
                                   .$chloride_twa_48>98&.$chloride_twa_48<=100~ 4,
                                   .$chloride_twa_48>96&.$chloride_twa_48<=98 ~ 3,
                                   .$chloride_twa_48>94&.$chloride_twa_48<=96 ~ 2,
                                   .$chloride_twa_48<=94~1,
                                   TRUE ~ as.numeric(NA)),
         na_48_mean_cat=case_when(.$sodium_twa_48>145 ~ 2,
                                .$sodium_twa_48<=145  & .$sodium_twa_48>=135 ~ 0,
                                .$sodium_twa_48<135 ~ 1,
                                TRUE ~ as.numeric(NA)),
         na_48_mean_catd=case_when(.$sodium_twa_48>146 ~ 11,
                                   .$sodium_twa_48>144&.$sodium_twa_48<=146 ~ 10,
                                   .$sodium_twa_48>142&.$sodium_twa_48<=144 ~ 9,
                                   .$sodium_twa_48>140&.$sodium_twa_48<=142 ~ 8,
                                   .$sodium_twa_48>138&.$sodium_twa_48<=140 ~ 7,
                                   .$sodium_twa_48>136&.$sodium_twa_48<=138 ~ 6,
                                   .$sodium_twa_48>134&.$sodium_twa_48<=136 ~ 5,
                                   .$sodium_twa_48>132&.$sodium_twa_48<=134 ~ 4,
                                   .$sodium_twa_48>130&.$sodium_twa_48<=132 ~ 3,
                                   .$chloride_twa_48<=130~1,
                                   TRUE ~ as.numeric(NA)),
         sid_48_mean_cat=case_when(.$SID_twa_48>37 ~ 2,
                                .$SID_twa_48<=37  & .$SID_twa_48>=31 ~ 0,
                                .$SID_twa_48<31 ~ 1,
                                TRUE ~ as.numeric(NA)),
         sid_48_mean_catd=case_when(.$SID_twa_48>44 ~ 11,
                                .$SID_twa_48>42  & .$SID_twa_48<=44 ~ 10,
                                .$SID_twa_48>40  & .$SID_twa_48<=42 ~ 9,
                                .$SID_twa_48>38  & .$SID_twa_48<=40 ~ 8,
                                .$SID_twa_48>36  & .$SID_twa_48<=38 ~ 7,
                                .$SID_twa_48>34  & .$SID_twa_48<=36 ~ 6,
                                .$SID_twa_48>32  & .$SID_twa_48<=34 ~ 5,
                                .$SID_twa_48>30  & .$SID_twa_48<=32 ~ 4,
                                .$SID_twa_48>28  & .$SID_twa_48<=30 ~ 3,
                                .$SID_twa_48>26  & .$SID_twa_48<=28 ~ 2,
                                .$SID_twa_48<=26 ~ 1,
                                TRUE ~ as.numeric(NA)),
         cl_48_mean_catq=cut(chloride_twa_48,breaks=10,labels=c("1","2","3","4","5","6","7","8","9","10")),
         delta_cl_d2=chloride_twa_48-chloride_first,
         delta_na_d2=sodium_twa_48-sodium_first,
         delta_sid_d2=SID_twa_48-SID_first
         
         
         )%>%
   mutate_at(c("gender","pasthistory_chf","pasthistory_pvd","pasthistory_hypertension","pasthistory_copd","pasthistory_diabetes","cl_adm_cat","na_adm_cat","sid_adm_cat"
               
               ,"cl_24_min_cat","cl_24_max_cat","na_24_min_cat","na_24_max_cat","sid_24_min_cat","sid_24_max_cat","ethnicity","cl_24_mean_cat","na_24_mean_cat","sid_24_mean_cat","surgery","sid_24_mean_catd","cl_24_mean_catd","na_24_mean_catd"

               ,"cl_48_min_cat","cl_48_max_cat","na_48_min_cat","na_48_max_cat","sid_48_min_cat","sid_48_max_cat","cl_48_mean_cat","na_48_mean_cat","sid_48_mean_cat","surgery","sid_48_mean_catd","cl_48_mean_catd","na_48_mean_catd"               
               ),as.factor)

#Focusing on medical patients
df_me<-df_2%>%filter(surgery==0)
```

## Dataset report

```{r}
view(dfSummary(df_me))
```

# Selecting variables

Main predictors: (you may want to use 1 OR 2 in a model)
1. Rough categories: cl_24_mean_cat,na_24_mean_cat,sid_24_mean_cat
2. Detailed categories: cl_24_mean_catd,na_24_mean_catd,sid_24_mean_catd

Patients’ baseline characteristics: (we may want to use all of the variables) age_fixed,gender,ethnicity,pasthistory_chf,pasthistory_pvd,pasthistory_hypertension,pasthistory_copd,pasthistory_diabetes,pasthistory_cld,pasthistory_malignancy,creat1, potassium_adm, bicarbonate_adm, ua_adm, heartrate_adm, systolic_adm, temperature_adm

Variables for the first 24 Horus: (we may want to use all of the variables) 
_I have used the mean_
glucose_mean_day1, potassium_mean_day1, bicarbonate_mean_day1, albumin_mean_day1, hemoglobin_mean_day1, wbc_mean_day1, platelets_mean_day1, heartrate_mean_day1, systolic_mean_day1, temperature_mean_day1


```{r}
## Excluding not desired variables
df_me_selected<-df_me%>%select( -one_of (c('apacheadmissiondx' ,'unitadmitsource' ,'BMI_group' ,'unitType' ,'hospitalAdmitSource' ,'hospLOS_prior_ICUadm_days' ,'unitlos_days' ,'hosplos_days' ,'ventdays_days' ,'hospitalmortality' ,'icumortality' ,'urineoutput_total_24' ,'urineoutput_total_48' ,'creat1offset' ,'first_rrtoffset' ,'rrt_bin' ,'peakcreat48h_offset' ,'peakcreat48h' ,'peakcreat7d' ,'peakcreat7d_offset' ,'peakcreat7d_to_discharge_offsetgap'  ,'cl_adm_cat' ,'na_adm_cat' ,'sid_adm_cat' ,'cl_24_min_cat' ,'cl_24_max_cat' ,'na_24_min_cat' ,'na_24_max_cat' ,'sid_24_min_cat' ,'sid_24_max_cat' ,'cl_24_mean_cat' ,'cl_24_mean_catd' ,'na_24_mean_cat' ,'na_24_mean_catd' ,'sid_24_mean_cat' ,'sid_24_mean_catd' ,'cl_24_mean_catq' ,'delta_cl_d1' ,'delta_na_d1' ,'delta_sid_d1','AKIstage')) )
```

# Imputation on selected variables

## Mean Imputation

```{r}
df_me_selected_imputed<-df_me_selected

for(i in 1:ncol(df_me_selected_imputed)){
  df_me_selected_imputed[is.na(df_me_selected_imputed[,i]), i] <- mean(df_me_selected_imputed[,i], na.rm = TRUE)
}
```

## Recovering non-imputable columns

We are not imputing variables that: 

A) *Almost sure they are MNAR* 
fio2_first, peep_first, peep_min_48, 

B) *Could be MNAR*
pao2_first, paco2_first, ph_first, anoingap_first, basedeficit_first, baseexcess_first,

fio2_min_24, pao2_min_24, pH_min_24, aniongap_min_24, basedeficit_min_24, baseexcess_min_24, peep_min24, fio2_max_24, pao2_max_24, paco2_max_24, pH_max_24, aniongap_max_24, basedeficit_max_24, baseexcess_max_24, peep_max24, 

fio2_min_48, pao2_min_48, paco2_min_48, pH_min_48, aniongap_min_48, basedeficit_min_48, baseexcess_min_48, fio2_max_48, pao2_max_48, paco2_max_48, pH_max_48, aniongap_max_48, basedeficit_max_48, baseexcess_max_48, peep_max_48

We are also not imputing those variables with more than 70% missing values.

### Studying variables missing values

```{r}
missing_values<-as.data.frame(sapply(df_me_selected, function(x) round(sum(is.na(x))/length(x),2)))
names(missing_values)<-'percentage_missing'

toomany_missing<-rownames(subset(missing_values,missing_values$percentage_missing>0.70))
```

### Final replacement

We are not imputing values listed above, so we replace the imputed columns with the orginal ones

```{r}
keep.origninal<-select(df_me_selected,c(
# MNAR  
fio2_first, peep_first, peep_min_48, 
pao2_first, paco2_first, pH_first, aniongap_first, basedeficit_first, baseexcess_first,

fio2_min_24, pao2_min_24, pH_min_24, aniongap_min_24, basedeficit_min_24, baseexcess_min_24, peep_min24, fio2_max_24, pao2_max_24, paco2_max_24, pH_max_24, aniongap_max_24, basedeficit_max_24, baseexcess_max_24, peep_max24, 

fio2_min_48, pao2_min_48, paco2_min_48, pH_min_48, aniongap_min_48, basedeficit_min_48, baseexcess_min_48, fio2_max_48, pao2_max_48, paco2_max_48, pH_max_48, aniongap_max_48, basedeficit_max_48, baseexcess_max_48, peep_max_48,
# too many missing values
fio2_first,pao2_first,paco2_first,pH_first,basedeficit_first,baseexcess_first,peep_first,fio2_min_24,basedeficit_min_24,peep_min24,fio2_min_48,basedeficit_min_48,peep_min_48,fio2_max_24,basedeficit_max_24,peep_max24,fio2_max_48,basedeficit_max_48,peep_max_48,albumin_first,bilirubin_first,wbc_first,platelets_first,ua_adm,

# we don't want to impute categorical variables
aki,

# concentration variables (MAYBE WE WANT TO IMPUTE THEM IN THE FUTURE)

total_na_meq_day1, total_cl_meq_day1, total_volume_ml_day1

)
)

# Final replacement

df_me_selected_imputed[,names(keep.origninal)] <- keep.origninal
```

## Multiple imputation

The following code has not been finished yet:

```{r eval=FALSE, include=FALSE}
#min and max possible values need to be set per variable.
bds <- matrix(c(7, 0, 100), nrow = 1, ncol = 3)# we need a matrix of bounds(number of column to bound, min, max)
relevant_data_imputation_matrix<-amelia(df_me_selected, parallel = "multicore",idvars = 'Case_Name',noms = c('Gender','Category','Q_AUT50MAP_total_min','Q_AUT65MAP_total_min','Q_AOT65MAP_total_min','Q_total_pressors_formula_mg')
                               #, ords = #Ords is only for ingegers, not floats
                               ,bounds = bds
                               )
FinalHemodynamicsDataset<-df_me_selected$imputations$imp5
names(df_me_selected)[names(df_me_selected)=="STS_RiskAlgorithm"]<-"STS_RiskAlgorithm_multiple_imp"
summary(df_me_selected$STS_RiskAlgorithm_multiple_imp)
#STS risk cannot be negative
# to avoid problems when using log
#there can not be 0 values in a log(variable) in a logreg
```

# Dataset report

```{r}
view(dfSummary(df_me_selected_imputed))
```

# Data export

```{r}
write.csv(df_me_selected_imputed,'df_me_selected_imputed.csv', row.names = F)
```
